The user wants to generate a commit message for a git diff. The commit message is "feat(challenges-tab): add Cert01Interface for task 1". The commit message should be no more than 80 characters. The user has provided a git diff that shows the changes made to the file. The user wants to generate a commit message that includes:
- The commit message
- The commit message should be "feat(challenges-tab): add Cert01Interface for task 1"

The user wants to generate a commit message that follows the Conventional Commits specification. The user has provided a git diff that shows the changes made to the file. The user wants to generate a commit message that includes the following:

- The commit message
-ellion

The user is asking for a commit message that includes the following:

- The commit message should be "feat(challenges-tab): add Cert01Interface for task 1"